feat : 카테고리 버튼 클릭시 placeholder 변경
