+++ src/pages/Search.tsx
@@ -3,15 +3,7 @@ import styled from 'styled-components';
 
 import background from '../asset/image/searchHomeBackground.svg';
 import Hover from '../components/common/Hover';
-import {
-  CategoryBar,
-  CategoryButton,
-  HrContainer,
-  Preview,
-  Searchbar,
-  TitleBoard,
-  TransparentHeader,
-} from '../components/search';
+import { CategoryButton, HrContainer, Preview, Searchbar, TitleBoard, TransparentHeader } from '../components/search';
 
 const Search = () => {
   const [isHover, setIsHover] = useState(false);
@@ -31,7 +23,6 @@ const Search = () => {
         <HrContainer />
       </Background>
       <Searchbar />
-      <CategoryBar />
       <StContentSection onMouseOver={handleHover} onMouseOut={handleHoverOut}>
         {[1, 2, 3, 4, 5, 6, 6, 7, 7, 7, 7, 7, 7, 7, 2, 2, 2, , 2, , 2, 2, , 2].map((_, idx) => (
           <Preview key={idx} profileImg="" name="Wedge Studio" recommandCount={129} visibleCount={129} />
@@ -53,5 +44,5 @@ const Background = styled.section`
 const StContentSection = styled.section`
   display: grid;
   grid-template-columns: repeat(5, 1fr);
-  margin: 1.185rem 1.875rem;
+  margin: 1.8125rem 1.875rem;
 `;
